refactor(core): 🔨 split additional verifiers
