+++ DomainDetective/DomainHealthCheck.Verification.EdnsSupport.cs
@@ -1,0 +1,23 @@
+using DnsClientX;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DomainDetective {
+    public partial class DomainHealthCheck {
+        /// <summary>
+        /// Tests authoritative name servers for EDNS support.
+        /// </summary>
+        /// <param name="domainName">Domain to verify.</param>
+        /// <param name="cancellationToken">Token to cancel the operation.</param>
+        public async Task VerifyEdnsSupport(string domainName, CancellationToken cancellationToken = default) {
+            cancellationToken.ThrowIfCancellationRequested();
+            if (string.IsNullOrWhiteSpace(domainName)) {
+                throw new ArgumentNullException(nameof(domainName));
+            }
+            domainName = NormalizeDomain(domainName);
+            UpdateIsPublicSuffix(domainName);
+            await EdnsSupportAnalysis.Analyze(domainName, _logger);
+        }
+    }
+}
